+++ CHANGELOG.md
@@ -4,6 +4,7 @@
   - Display 24 hour usage
   - Add button for displaying usage by period
   - Record the bundle data on manual refresh in addition to displaying it
+  - Add button for opening mobile data settings screen
 
 + Fixes: